requirements.txt
torch
transformers>=4.44.0
decord
av
opencv-python
tqdm
pandas
numpy
seaborn
matplotlib
wandb
pyyaml
